+++ sim/paladin/talents_protection.go
@@ -58,7 +58,7 @@ func (paladin *Paladin) applyHallowedGround() {
 	paladin.AddStaticMod(core.SpellModConfig{
 		ClassMask: SpellMaskConsecration,
 		Kind:      core.SpellMod_PowerCost_Pct,
-		IntValue:  -(40 * int64(paladin.Talents.HallowedGround)),
+		IntValue:  -40 * paladin.Talents.HallowedGround,
 	})
 }
